+++ android/src/main/kotlin/com/soriana/zebra_bt_printer/bridge/PrinterExceptions.kt
@@ -1,0 +1,7 @@
+package com.soriana.zebra_bt_printer.bridge
+
+internal class PaperOutException(message: String) : Exception(message)
+
+internal class PrintTimeoutException(message: String) : Exception(message)
+
+internal class CalibrateTimeoutException(message: String) : Exception(message)